+++ src/assets/scripts/client/aircraft/AircraftCommander.js
@@ -30,9 +30,10 @@ import { radiansToDegrees } from '../utilities/unitConverters';
  * @class AircraftCommander
  */
 export default class AircraftCommander {
-    constructor(onChangeTransponderCode) {
+    constructor(aircraftController) {
         this._eventBus = EventBus;
-        this._onChangeTransponderCode = onChangeTransponderCode;
+        this._aircraftController = aircraftController;
+        this._onChangeTransponderCode = aircraftController.onRequestToChangeTransponderCode;
     }
 
     /**
@@ -82,6 +83,7 @@ export default class AircraftCommander {
                 response.push(retval[1]);
 
                 if (retval[2]) {
+                    // eslint-disable-next-line prefer-destructuring
                     response_end = retval[2];
                 }
             }
@@ -335,10 +337,10 @@ export default class AircraftCommander {
         if (_isNil(runwayModel)) {
             const previousRunwayModel = aircraft.fms.arrivalRunwayModel;
             const readback = {};
-            readback.log = `unable to find Runway ${runwayName} on our charts, ` +
-                `expecting Runway ${previousRunwayModel.name} instead`;
-            readback.say = `unable to find Runway ${radio_runway(runwayName)} on our ` +
-                `charts, expecting Runway ${previousRunwayModel.getRadioName()} instead`;
+            readback.log = `unable to find Runway ${runwayName} on our charts, `
+                + `expecting Runway ${previousRunwayModel.name} instead`;
+            readback.say = `unable to find Runway ${radio_runway(runwayName)} on our `
+                + `charts, expecting Runway ${previousRunwayModel.getRadioName()} instead`;
 
             return [false, readback];
         }
@@ -603,7 +605,7 @@ export default class AircraftCommander {
         const runway = aircraft.fms.departureRunwayModel;
         const spotInQueue = runway.getAircraftQueuePosition(aircraft.id);
         const isInQueue = spotInQueue > -1;
-        const aircraftAhead = runway.queue[spotInQueue - 1];
+        const aircraftAhead = this._aircraftController.findAircraftById(runway.queue[spotInQueue - 1]);
         const wind = airport.getWind();
         const roundedWindAngleInDegrees = round(radiansToDegrees(wind.angle) / 10) * 10;
         const roundedWindSpeed = round(wind.speed);
@@ -645,10 +647,10 @@ export default class AircraftCommander {
 
         // see #1154, we may have been rerouted since we started taxiing.
         if (!aircraft.fms.isRunwayModelValidForSid(runway)) {
-            readback.log = `according to our charts, Runway ${runway.name} ` +
-                `is not valid for the ${aircraft.fms.getSidIcao()} departure`;
-            readback.say = `according to our charts, Runway ${runway.getRadioName()} ` +
-                `is not valid for the ${aircraft.fms.getSidName()} departure`;
+            readback.log = `according to our charts, Runway ${runway.name} `
+                + `is not valid for the ${aircraft.fms.getSidIcao()} departure`;
+            readback.say = `according to our charts, Runway ${runway.getRadioName()} `
+                + `is not valid for the ${aircraft.fms.getSidName()} departure`;
 
             return [false, readback];
         }
@@ -659,12 +661,13 @@ export default class AircraftCommander {
         aircraft.setFlightPhase(FLIGHT_PHASE.TAKEOFF);
         aircraft.scoreWind('taking off');
 
-        readback.log = `wind ${roundedWindAngleInDegrees} at ${roundedWindSpeed}, Runway ${runway.name}, ` +
-            'cleared for takeoff';
+        readback.log = `wind ${roundedWindAngleInDegrees} at ${roundedWindSpeed}, `
+            + `Runway ${runway.name}, cleared for takeoff`;
 
         // We have to make it say winned to make it sound like "Wind" and not "Whined"
-        readback.say = `winned ${radio_spellOut(roundedWindAngleInDegrees)} at ` +
-            `${radio_spellOut(roundedWindSpeed)}, Runway ${radio_runway(runway.name)}, cleared for takeoff`;
+        readback.say = `winned ${radio_spellOut(roundedWindAngleInDegrees)} at `
+            + `${radio_spellOut(roundedWindSpeed)}, Runway ${radio_runway(runway.name)}, `
+            + 'cleared for takeoff';
 
         return [true, readback];
     }